+++ modules/core/src/main/scala/codec/NumericCodecs.scala
@@ -6,7 +6,6 @@ package skunk
 package codec
 
 import skunk.data.Type
-import skunk.util.ArrParser
 
 trait NumericCodecs {
 
@@ -24,7 +23,13 @@ trait NumericCodecs {
   val float4: Codec[Float]   = Codec.simple(_.toString, safe(_.toFloat), Type.float4)
   val float8: Codec[Double]  = Codec.simple(_.toString, safe(_.toDouble), Type.float8)
 
-  val _int4: Codec[Arr[Int]] = Codec.simple(_.encode(_.toString), ArrParser.parseWith(safe(_.toInt)), Type._int4)
+  // Array types
+  val _int2: Codec[Arr[Short]] = Codec.array(_.toString, safe(_.toShort), Type._int2)
+  val _int4: Codec[Arr[Int]]   = Codec.array(_.toString, safe(_.toInt),   Type._int4)
+  val _int8: Codec[Arr[Long]]  = Codec.array(_.toString, safe(_.toLong),  Type._int8)
+  val _numeric: Codec[Arr[BigDecimal]] = Codec.array(_.toString, safe(BigDecimal(_)), Type._numeric)
+  val _float4: Codec[Float]   = Codec.simple(_.toString, safe(_.toFloat), Type._float4)
+  val _float8: Codec[Double]  = Codec.simple(_.toString, safe(_.toDouble), Type._float8)
 
 }
